requirements.txt
keras==2.8.0
matplotlib==3.5.0
numpy==1.21.2
opencv_python==4.5.4.60
pandas==1.3.4
Pillow==9.0.1
scikit_learn==1.0.2
seaborn==0.11.2
tensorflow==2.7.0
